@@ -3,11 +3,12 @@
 namespace App;
 
 use Carbon\Carbon;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Reply extends Model
 {
-    use Favoritable, RecordsActivity;
+    use HasFactory, Favoritable, RecordsActivity;
     protected $guarded=[];
     protected $with = ['owner', 'favorites', 'thread'];
     protected $appends = ['favoritesCount', 'isFavorited'];
