@@ -1,9 +1,9 @@
 <?php
 function create($class, $attributes=[], $items = null)
 {
-    return factory($class, $items)->create($attributes);
+    return $class::factory($items)->create($attributes);
 }
 function make($class, $attributes=[], $items = null)
 {
-    return factory($class,$items)->make($attributes);
+    return $class::factory($items)->make($attributes);
 }
